+++ src/Console/AppInstall.php
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-namespace GrahamCampbell\Core\Commands;
+namespace GrahamCampbell\Core\Console;
 
 use Illuminate\Console\Command;
 use Illuminate\Contracts\Events\Dispatcher;